+++ app/src/main/java/dev/shipi/mylittlespiders/components/forms/Input.kt
@@ -1,4 +1,4 @@
-package dev.shipi.mylittlespiders.lib.presentation
+package dev.shipi.mylittlespiders.components.forms
 
 import java.time.LocalDate
