+++ app/src/main/java/dev/shipi/mylittlespiders/components/screens/ErrorScreen.kt
@@ -1,4 +1,4 @@
-package dev.shipi.mylittlespiders.components
+package dev.shipi.mylittlespiders.components.screens
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -6,8 +6,10 @@ import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.width
-import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.rounded.Warning
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
 import androidx.compose.material3.LargeTopAppBar
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -19,12 +21,12 @@ import androidx.compose.ui.unit.dp
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
-fun LoadingScreen() {
+fun ErrorScreen(message: String? = null) {
     Column {
         LargeTopAppBar(
             title = {
                 Text(
-                    text = "Loading...",
+                    text = "Oh no!",
                     style = MaterialTheme.typography.headlineMedium
                 )
             },
@@ -36,15 +38,38 @@ fun LoadingScreen() {
             verticalArrangement = Arrangement.Center,
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            CircularProgressIndicator(modifier = Modifier
-                .height(64.dp)
-                .width(64.dp))
+            Column {
+                Icon(
+                    modifier = Modifier
+                        .height(128.dp)
+                        .width(128.dp),
+                    imageVector = Icons.Rounded.Warning,
+                    contentDescription = "Warning symbol",
+                    tint = MaterialTheme.colorScheme.error
+                )
+                Text(
+                    text = "Something went wrong",
+                    style = MaterialTheme.typography.headlineSmall
+                )
+                if (message != null) {
+                    Text(
+                        text = message,
+                        style = MaterialTheme.typography.bodyLarge
+                    )
+                }
+            }
         }
     }
 }
 
 @Preview(showBackground = true, showSystemUi = true)
 @Composable
-fun LoadingScreenPreview() {
-    LoadingScreen()
+fun ErrorScreenWithMessagePreview() {
+    ErrorScreen("Reason")
+}
+
+@Preview(showBackground = true, showSystemUi = true)
+@Composable
+fun ErrorScreenNoMessagePreview() {
+    ErrorScreen()
 }
