+++ app/src/main/java/dev/shipi/mylittlespiders/presentation/details/DetailsScreen.kt
@@ -30,12 +30,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
-import dev.shipi.mylittlespiders.components.ErrorScreen
-import dev.shipi.mylittlespiders.components.LoadingScreen
-import dev.shipi.mylittlespiders.components.NetworkNotAvailableWidget
+import dev.shipi.mylittlespiders.components.screens.ErrorScreen
+import dev.shipi.mylittlespiders.components.screens.LoadingScreen
+import dev.shipi.mylittlespiders.components.widgets.NetworkNotAvailableWidget
 import dev.shipi.mylittlespiders.domain.model.Entry
 import dev.shipi.mylittlespiders.domain.model.FriendDetails
-import dev.shipi.mylittlespiders.lib.presentation.ViewState
+import dev.shipi.mylittlespiders.components.ViewState
 import dev.shipi.mylittlespiders.presentation.details.view.EntryRow
 import java.lang.Exception
 import java.time.LocalDate
