+++ app/src/main/java/dev/shipi/mylittlespiders/presentation/entry/EntryFormViewModel.kt
@@ -2,7 +2,7 @@ package dev.shipi.mylittlespiders.presentation.entry
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import dev.shipi.mylittlespiders.lib.presentation.Input
+import dev.shipi.mylittlespiders.components.forms.Input
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
