+++ app/src/main/java/dev/shipi/mylittlespiders/presentation/entry/add/AddEntryViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
 import dev.shipi.mylittlespiders.domain.model.NewEntry
 import dev.shipi.mylittlespiders.domain.usecase.AddEntry
-import dev.shipi.mylittlespiders.lib.presentation.ViewState
+import dev.shipi.mylittlespiders.components.ViewState
 import dev.shipi.mylittlespiders.presentation.entry.EntryFormViewModel
 import dev.shipi.mylittlespiders.services.NetworkObserver
 import kotlinx.coroutines.flow.MutableStateFlow
