+++ app/src/main/java/dev/shipi/mylittlespiders/presentation/list/ListViewModel.kt
@@ -7,7 +7,7 @@ import dev.shipi.mylittlespiders.domain.model.Friend
 import dev.shipi.mylittlespiders.domain.usecase.DeleteFriend
 import dev.shipi.mylittlespiders.domain.usecase.GetFriendList
 import dev.shipi.mylittlespiders.domain.usecase.RefreshFriendList
-import dev.shipi.mylittlespiders.lib.presentation.ViewState
+import dev.shipi.mylittlespiders.components.ViewState
 import dev.shipi.mylittlespiders.services.NetworkObserver
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
